requirements.txt
google-genai>=1.0.0
resend>=2.0.0
feedparser>=6.0.10
requests>=2.31.0
pyyaml>=6.0
python-dotenv>=1.0.0
jinja2>=3.1.0
